Don't generate races when recording results
This is a workaround for: the recordRaceResult action needs to resolve
before we do generateRace, AND the selectors need to refresh, but that
will not happen by the time we also dispatch the generateRaces action
from inside the recordResult function. So: just don't generate new races
after recording results. Probably thunks or something could solve this
properly, but for now this is good enough.